+++ linbox/algorithms/polynomial-matrix/polynomial-fft-transform.h
@@ -160,7 +160,7 @@ namespace LinBox {
 			chrono.start();
 			
 			// compute w^(-1) mod p = w^(2^lpts - 1)
-			_invw = Givaro::powmod(_w, (1UL<<ln) - 1, _pl);
+			_invw = Givaro::powmod(_w, ((uint64_t)1<<ln) - 1, _pl);
 			
 			size_t pos = 0;
 			//uint64_t wi = 1;
